Update Mastermind.c
Changes made by Prajna and Pragya:
We have added another difficulty level called Standard between Novice and Expert. 
We have restricted the number of tries allowed to 15, 20 and 25 respectively based on the difficulty level. 
We have added a 5th option in the beginning of the game called Quit, if the user wants to quit the game.
We have put the entire game within a loop which can be exited using Quit option.