Add files via upload
This is the final version of our Project - Mastermind. The features include playing the game as code breaker, playing as code maker, choosing difficulty level, reading about the game, reading how to play the game, simulating the game and option to quit. 
The most recent change is playing as code maker.